requirements.txt
certifi==2018.4.16
future==0.16.0
peewee==3.6.4
pkg-resources==0.0.0
python-telegram-bot==10.1.0
